# api-platform/infra/openapi.yaml
...
openapi: "3.0.1"
info:
  title: City of Alpenglow Developer API
  version: "2.1.0"
  description: >-
    Public REST API for the fictional City of Alpenglow, CO: permit records,
    business licenses, and parks & facilities, each served by its own
    microservice. v2 is current; v1 still works but answers with
    Deprecation/Sunset headers pointing at its successor. All endpoints except
    /v2/status and /v1/ping require an x-api-key header; the shared demo key
    is printed at https://api.demos.planetek.org.
  contact:
    name: Planetek
    url: https://demos.planetek.org
...
paths:
...
  /v1/permits:
    get:
      tags: [permits]
      summary: List permits (deprecated)
      deprecated: true
      description: Bare-array response, page capped at 50, no cursor. Replies carry Deprecation, Sunset, and Link successor-version headers. Use GET /v2/permits.
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - { name: type, in: query, required: false, schema: { $ref: "#/components/schemas/PermitType" } }
        - { name: status, in: query, required: false, schema: { $ref: "#/components/schemas/PermitStatus" } }
      responses:
        "200":
          description: Up to 50 permits, no pagination.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Permit" }
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri: ${permits_uri}
  /v1/permits/{id}:
    get:
      tags: [permits]
      summary: Get a permit (deprecated)
      deprecated: true
      description: Bare-object response with deprecation headers. Use GET /v2/permits/{id}.
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, example: PRM-2025-0104 }
      responses:
        "200":
          description: The permit.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Permit" }
        "404": { $ref: "#/components/responses/NotFound" }
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri: ${permits_uri}
...
  /v2/permits:
    get:
      tags: [permits]
      summary: List permits
      description: Cursor-paginated envelope. Filter with ?type= and ?status=; page with ?limit= (max 50) and the opaque meta.nextToken from the previous page.
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - { name: type, in: query, required: false, schema: { $ref: "#/components/schemas/PermitType" } }
        - { name: status, in: query, required: false, schema: { $ref: "#/components/schemas/PermitStatus" } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
        - { name: nextToken, in: query, required: false, schema: { type: string }, description: "Opaque cursor from the previous page's meta.nextToken." }
      responses:
        "200":
          description: One page of permits.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Permit" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri: ${permits_uri}
  /v2/permits/{id}:
    get:
      tags: [permits]
      summary: Get a permit
      description: Responses carry a strong ETag; repeat the call with If-None-Match and an unchanged record answers 304 with no body.
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, example: PRM-2025-0104 }
        - { name: If-None-Match, in: header, required: false, schema: { type: string }, description: "ETag from a previous response; matching answers 304." }
      responses:
        "200":
          description: The permit, enveloped.
          headers:
            ETag:
              schema: { type: string }
              description: Strong validator for If-None-Match on later calls.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Permit" }
        "304": { $ref: "#/components/responses/NotModified" }
        "404": { $ref: "#/components/responses/NotFound" }
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri: ${permits_uri}
  /v2/permits/{id}/inspections:
    get:
      tags: [permits]
      summary: List inspection requests for a permit
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, example: PRM-2025-0104 }
      responses:
        "200":
          description: Inspection requests currently on file (visitor-created ones expire after 24h).
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Inspection" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
        "404": { $ref: "#/components/responses/NotFound" }
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri: ${permits_uri}
    post:
      tags: [permits]
      summary: Request an inspection
      description: >-
        The write surface, and the request-validation exhibit. API Gateway
        checks the body against the InspectionRequest model (required fields,
        enums, formats, no unknown properties) and rejects invalid payloads
        with a 400 before any Lambda runs. Created records expire
        automatically after 24 hours. Only approved or issued permits accept
        inspection requests (409 otherwise). Send an Idempotency-Key header
        and retrying the identical request within 24 hours replays the
        original 201 (marked Idempotency-Replayed: true) instead of creating
        a duplicate.
      security: [{ ApiKeyAuth: [] }]
      x-amazon-apigateway-request-validator: validate-body
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, example: PRM-2025-0104 }
        - { name: Idempotency-Key, in: header, required: false, schema: { type: string, pattern: '^[A-Za-z0-9_-]{1,64}$' }, description: "Optional client-chosen key (1-64 chars) that makes this POST safely retryable for 24 hours." }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InspectionRequest" }
      responses:
        "201":
          description: Inspection request created (expires in 24h). A replayed response carries Idempotency-Replayed true.
          headers:
            Idempotency-Replayed:
              schema: { type: string }
              description: Present (true) when this 201 is the stored original, replayed for a repeated Idempotency-Key.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Inspection" }
        "400":
          description: Body failed gateway schema validation. The Lambda never ran.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Permit exists but is not in an inspectable status.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri: ${permits_uri}
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: x-api-key
      in: header
      description: Usage-plan API key. The shared demo-tier key is printed on the docs page; partner-tier keys are issued per integrator.
  responses:
    NotFound:
      description: No record with that id.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotModified:
      description: Not modified — the ETag presented in If-None-Match still matches, so no body is sent.
  schemas:
    PermitType:
      type: string
      enum: [building, electrical, plumbing, mechanical, sign, fence, solar, event]
    PermitStatus:
      type: string
      enum: [submitted, under-review, approved, issued, denied, closed]
...
    Permit:
      type: object
      properties:
        id: { type: string, example: PRM-2025-0104 }
        type: { $ref: "#/components/schemas/PermitType" }
        status: { $ref: "#/components/schemas/PermitStatus" }
        description: { type: string, example: 200A service upgrade }
        address: { type: string, example: 1420 Larkspur Ave }
        applicant: { type: string, example: Peak & Pine Electric }
        valuation: { type: integer, description: "USD", example: 12400 }
        submittedAt: { type: string, example: "2025-03-11" }
...
    InspectionRequest:
      type: object
      description: Enforced by API Gateway's request validator, not by application code.
      required: [type, preferredDate, contactEmail]
      additionalProperties: false
      properties:
        type:
          type: string
          enum: [footing, rough, electrical-service, plumbing-top-out, final]
        preferredDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
          description: ISO date (YYYY-MM-DD).
        contactEmail:
          type: string
          pattern: '^[^@\s]+@[^@\s]+\.[^@\s]+$'
        notes:
          type: string
          maxLength: 500
    Inspection:
      type: object
      properties:
        id: { type: string, example: INSP-LZK4A2M01 }
        permitId: { type: string, example: PRM-2025-0104 }
        type: { type: string, example: rough }
        preferredDate: { type: string, example: "2026-08-01" }
        contactEmail: { type: string, example: inspector@example.com }
        notes: { type: string }
        status: { type: string, example: requested }
        requestedAt: { type: string, format: date-time }
...
    PageMeta:
      type: object
      properties:
        count: { type: integer, description: Items in this page. }
        nextToken: { type: string, description: Opaque cursor for the next page; absent on the last page. }
...
    Error:
      type: object
      properties:
        error: { type: string, example: not_found }
        message: { type: string }